+++ test/cell_multi_cell_refactor/cell.py
@@ -1,5 +1,3 @@
-"""metadata_test.py"""
-
 import unittest
 import sys
 import os
@@ -10,15 +8,14 @@ sys.path.insert(
 )
 
 import fpdf
-from test.utilities import relative_path_to, set_doc_date_0, calculate_hash_of_file
+from test.utilities import assert_pdf_equal
 
-# python -m unittest test.cell_multi_cell_refactor.cell.CRefactorTest
+# python -m unittest test.cell_multi_cell_refactor.cell
 
 
 class CRefactorTest(unittest.TestCase):
     def test_ln_positioning_and_page_breaking_for_cell(self):
         doc = fpdf.FPDF(format="letter", unit="pt")
-        set_doc_date_0(doc)
         doc.add_page()
 
         my_text_size = 36
@@ -56,12 +53,13 @@ class CRefactorTest(unittest.TestCase):
                 link="",
             )
 
-        outfile = relative_path_to("output.pdf")
-        doc.output(outfile)
-        known_good_hash = "d49bb2b7c9b3504ef45309a4a155cfc2"
-        self.assertEqual(calculate_hash_of_file(outfile), known_good_hash)
-        os.unlink(outfile)
+        assert_pdf_equal(
+            self, doc, "test_ln_positioning_and_page_breaking_for_cell.pdf"
+        )
 
+
+if __name__ == "__main__":
+    unittest.main()
 
 ## Code used to create this test
 # doc = fpdf.FPDF(format = 'letter', unit = 'pt')
@@ -90,11 +88,6 @@ class CRefactorTest(unittest.TestCase):
 #         'dolore Ut in sint sunt exercitation aliquip elit velit dolor nisi '
 #         '')*100
 
-# for i in range(20):
-#   doc.cell(w = 72, h = my_text_size * 1.5, border = 1, ln = 2,
-#            txt = text[i * 100:i * 100 + 99], fill = 0, link = '')
 
-# outfile = relative_path_to('output.pdf')
-# doc.output(outfile)
-# print(calculate_hash_of_file(outfile))
-# os.unlink(outfile)
+if __name__ == "__main__":
+    unittest.main()