+++ test/class_unit_test/shape_tests.py
@@ -10,8 +10,9 @@ sys.path.insert(
 )
 
 import fpdf
-import test
-from test.utilities import relative_path_to, set_doc_date_0, calculate_hash_of_file
+from test.utilities import assert_pdf_equal
+
+# python -m unittest test.class_unit_test.shape_tests
 
 
 def next_row(pdf):
@@ -24,60 +25,34 @@ margin = 10
 
 
 class EllipseTest(unittest.TestCase):
-    """ShapeWriterTest"""
-
-    PDFClass = fpdf.FPDF
-
     def test_ellipse_not_circle(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Styles
-        counter = 0
-        for style in ["", "F", "FD", "DF", None]:
-            counter += 1
+        for counter, style in enumerate(["", "F", "FD", "DF", None]):
             pdf.ellipse(x=pdf.get_x(), y=pdf.get_y(), w=size / 2, h=size, style=style)
             pdf.set_x(pdf.get_x() + (size / 2) + margin)
-
             if counter % 3 == 0:
                 next_row(pdf)
 
-        outfile = relative_path_to("output.pdf")
-        pdf.output(outfile)
-
-        known_good_hash = "169345cb25b662b236d35aa0b473092f"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_ellipse_not_circle.pdf")
 
     def test_ellipse_style(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Styles
-        counter = 0
-        for style in ["", "F", "FD", "DF", None]:
-            counter += 1
+        for counter, style in enumerate(["", "F", "FD", "DF", None]):
             pdf.ellipse(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style=style)
             pdf.set_x(pdf.get_x() + size + margin)
-
             if counter % 3 == 0:
                 next_row(pdf)
 
-        outfile = relative_path_to("output1.pdf")
-        pdf.output(outfile)
-
-        known_good_hash = "2f08ed8338d7d421fe2a286ef6c00daf"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_ellipse_style.pdf")
 
     def test_ellipse_line_width(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Line Width
         for line_width in [1, 2, 3]:
             pdf.set_line_width(line_width)
             pdf.ellipse(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style=None)
@@ -89,35 +64,22 @@ class EllipseTest(unittest.TestCase):
             pdf.set_x(pdf.get_x() + size + margin)
         pdf.set_line_width(0.2)  # reset
 
-        outfile = relative_path_to("output2.pdf")
-        pdf.output(outfile)
-
-        known_good_hash = "9151e507484e32ca1577c6002ccafada"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_ellipse_line_width.pdf")
 
     def test_ellipse_draw_color(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Colors
         pdf.set_line_width(0.5)
         for gray in [70, 140, 210]:
             pdf.set_draw_color(gray)
             pdf.ellipse(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style=None)
             pdf.set_x(pdf.get_x() + size + margin)
 
-        outfile = relative_path_to("output3.pdf")
-        pdf.output(outfile)
-
-        known_good_hash = "ad08d121648ee2b6e38982cdcce01688"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_ellipse_draw_color.pdf")
 
     def test_ellipse_fill_color(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
         pdf.set_fill_color(240)
@@ -125,72 +87,40 @@ class EllipseTest(unittest.TestCase):
             pdf.set_draw_color(*color)
             pdf.ellipse(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style="FD")
             pdf.set_x(pdf.get_x() + size + margin)
-
         next_row(pdf)
 
-        outfile = relative_path_to("output4.pdf")
-        pdf.output(outfile)
-
-        known_good_hash = "2719bf0278757bb684d5d8e6e9cea5f5"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_ellipse_fill_color.pdf")
 
 
 class RectangleTest(unittest.TestCase):
-    """RectangleTest"""
-
-    PDFClass = fpdf.FPDF
-
     def test_rect_not_square(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Styles
-        counter = 0
-        for style in ["", "F", "FD", "DF", None]:
-            counter += 1
+        for counter, style in enumerate(["", "F", "FD", "DF", None]):
             pdf.rect(x=pdf.get_x(), y=pdf.get_y(), w=size / 2, h=size, style=style)
             pdf.set_x(pdf.get_x() + (size / 2) + margin)
-
             if counter % 3 == 0:
                 next_row(pdf)
 
-        outfile = relative_path_to("output.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "462a76e02de625f3b3c70a1f8eef9ebc"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_rect_not_square.pdf")
 
     def test_rect_style(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Styles
-        counter = 0
-        for style in ["", "F", "FD", "DF", None]:
-            counter += 1
+        for counter, style in enumerate(["", "F", "FD", "DF", None]):
             pdf.rect(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style=style)
             pdf.set_x(pdf.get_x() + size + margin)
-
             if counter % 3 == 0:
                 next_row(pdf)
 
-        outfile = relative_path_to("output1.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "be8e25fb5f4d3b1822c969d1478cef86"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_rect_style.pdf")
 
     def test_rect_line_width(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
-        # Line Width
         for line_width in [1, 2, 3]:
             pdf.set_line_width(line_width)
             pdf.rect(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style=None)
@@ -202,16 +132,10 @@ class RectangleTest(unittest.TestCase):
             pdf.set_x(pdf.get_x() + size + margin)
         pdf.set_line_width(0.2)  # reset
 
-        outfile = relative_path_to("output2.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "b5b9d94230af38f7429ccd73d3e342bb"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_rect_line_width.pdf")
 
     def test_rect_draw_color(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
         # Colors
@@ -221,16 +145,10 @@ class RectangleTest(unittest.TestCase):
             pdf.rect(x=pdf.get_x(), y=pdf.get_y(), w=size, h=size, style=None)
             pdf.set_x(pdf.get_x() + size + margin)
 
-        outfile = relative_path_to("output3.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "ab22c2b23e19e09387da55fd534d4f4c"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_rect_draw_color.pdf")
 
     def test_rect_fill_color(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
         pdf.set_fill_color(240)
@@ -241,22 +159,12 @@ class RectangleTest(unittest.TestCase):
 
         next_row(pdf)
 
-        outfile = relative_path_to("output4.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "b3a94f3b3c0282dcbbb5f6127b3dfaab"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_rect_fill_color.pdf")
 
 
 class LineTest(unittest.TestCase):
-    """LineTest"""
-
-    PDFClass = fpdf.FPDF
-
     def test_line(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
         def draw_diagonal_line(pdf, x, y):
@@ -274,16 +182,10 @@ class LineTest(unittest.TestCase):
             pdf.set_x(pdf.get_x() + size + margin)
         next_row(pdf)
 
-        outfile = relative_path_to("line_output.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "684bb1210caf57a77021124e1b8a81ef"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_line.pdf")
 
     def test_dash(self):
-        pdf = self.PDFClass(unit="mm")
-        set_doc_date_0(pdf)
+        pdf = fpdf.FPDF(unit="mm")
         pdf.add_page()
 
         def draw_diagonal_dash(pdf, x, y, *a, **k):
@@ -319,15 +221,8 @@ class LineTest(unittest.TestCase):
         x, y = pdf.get_x(), pdf.get_y()
         pdf.dashed_line(x, y, x + 100, y + 80, 6, 17)
 
-        outfile = relative_path_to("line_output1.pdf")
-        pdf.output(outfile)
-        # print(calculate_hash_of_file(outfile))
-        known_good_hash = "4cf8faa9baf3f1835c03fa4ac1e6eb29"
-        self.assertEqual(known_good_hash, calculate_hash_of_file(outfile))
-        os.unlink(outfile)
+        assert_pdf_equal(self, pdf, "test_dash.pdf")
 
 
 if __name__ == "__main__":
     unittest.main()
-
-## Development of demo mostly done as written above.