+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 ### Deprecated
 - the `type` parameter of `FPDF.image` method
 - the `infile` parameter of `Template` constructor
+- the `dest` parameter of `Template.render` method
 
 ## [2.1.0] - 2020-12-07
 ### Added